small file refactoring
